no custom file descriptor + pass env to child
child_process.fork() does not support customFds (custom file
descriptors), at least with Node.JS v0.5.9. Child need to communicate
through an event channel which is already implemented in node-qunit, so
this part of the patch should be safe.

The next change is passing the parent process environment to the child
upon creation. This is required when using $NODE_PATH or you the child
will not be able to find the node modules it depends upon (for example
the 'underscore' module).